Oops, should not have changed rotation and tilt to doubles in _readAngles() on the last commit. These two are sent to us explicitly as 32-bit floats, and we memcpy the values in so we *must* have the type correct.
git-svn-id: http://svn/svn/hcr/trunk@1094 2e3f476d-dfe7-4101-bf40-6ec35b1080dd

+++ src/hcrgui/HcrGuiMainWindow.cpp
@@ -670,8 +670,11 @@ HcrGuiMainWindow::_logMessage(std::string message) {
 void
 HcrGuiMainWindow::_readAngles()
 {
-    double rotation = 0.0;
-    double tilt = 0.0;
+    // Note that these two *must* be 32-bit floats, since the data we memcpy
+    // into these variables from the UDP stream is 32-bit floats.
+    float rotation = 0.0;
+    float tilt = 0.0;
+
     while (_angleSocket.hasPendingDatagrams()) {
         QByteArray datagram;
         datagram.resize(_angleSocket.pendingDatagramSize());
